+++ R/package.R
@@ -8,6 +8,7 @@ utils::globalVariables(c(
   "marginaleffects_contrast_hi",
   "marginaleffects_contrast_lo",
   "..bycols",
+  "..by",
   "..to_keep",
   "..cols",
   "..idx_by",