+++ inst/tinytest/test-equivalence.R
@@ -105,7 +105,7 @@ expect_snapshot_print(cmp, "equivalence-avg_comparisons")
 
 # bug on with call and symbols
 mod <- lm(mpg ~ hp * vs, data = mtcars)
-x <- avg_slopes(mod, by = "vs", variables = "hp", hypothesis = "pairwise")
+x <- avg_slopes(mod, by = "vs", variables = "hp", hypothesis = ~pairwise)
 x <- hypotheses(x, equivalence = c(-.2, .2))
 expect_inherits(x, "hypotheses")
 
@@ -123,12 +123,13 @@ mm <- predictions(
     mod,
     newdata = datagrid(grid_type = "balanced"),
     by = "source",
-    hypothesis = "pairwise")
+    hypothesis = ~pairwise)
 
 e1 <- test(pa, delta = delta, adjust = "none", side = "nonsuperiority", df = Inf)
 e2 <- hypotheses(mm, equivalence = c(-delta, delta))
 e1 <- e1[order(e1$contrast), ]
 e2 <- e2[order(e2$term), ]
+e2 <- e2[e2$term %in% e1$contrast,]
 expect_equivalent(e1$z.ratio, e2$statistic.nonsup, tol = 1e-6)
 expect_equivalent(e1$p.value, e2$p.value.nonsup, tol = 1e-6)
 
@@ -136,6 +137,7 @@ e1 <- test(pa, delta = delta, adjust = "none", side = "noninferiority", df = Inf
 e2 <- hypotheses(mm, equivalence = c(-delta, delta))
 e1 <- e1[order(e1$contrast), ]
 e2 <- e2[order(e2$term), ]
+e2 <- e2[e2$term %in% e1$contrast,]
 expect_equivalent(e1$z.ratio, e2$statistic.noninf, tolerance = 1e-6)
 expect_equivalent(e1$p.value, e2$p.value.noninf, tolerance = 1e-6)
 
@@ -143,6 +145,7 @@ e1 <- test(pa, delta = delta, adjust = "none", df = Inf)
 e2 <- hypotheses(mm, equivalence = c(-delta, delta))
 e1 <- e1[order(e1$contrast), ]
 e2 <- e2[order(e2$term), ]
+e2 <- e2[e2$term %in% e1$contrast,]
 expect_equivalent(e1$p.value, e2$p.value.equiv, tolerance = 1e-6)