+++ inst/tinytest/test-pkg-ggeffects.R
@@ -19,17 +19,17 @@ d <- structure(list(var_binom = structure(1:2, levels = c("0", "1"
 -2L))
 
 p1 <- predictions(
-  m, newdata = d, by = "var_binom", hypothesis = "pairwise", type = "response", transform = exp
+  m, newdata = d, by = "var_binom", hypothesis = ~pairwise, type = "response", transform = exp
 )
 p2 <- predictions(
-  m, newdata = d, by = "var_binom", hypothesis = "pairwise", type = "invlink(link)", transform = exp
+  m, newdata = d, by = "var_binom", hypothesis = ~pairwise, type = "invlink(link)", transform = exp
 ) |> suppressWarnings()
 p3 <- predictions(
-  m, newdata = d, by = "var_binom", hypothesis = "pairwise", type = NULL, transform = exp
+  m, newdata = d, by = "var_binom", hypothesis = ~pairwise, type = NULL, transform = exp
 ) |> suppressWarnings()
 
 # values
-expect_equal(p1$estimate, 0.9867827, tolerance = 1e-5)
-expect_equal(p2$estimate, 0.9867827, tolerance = 1e-5)
-expect_equal(p3$estimate, 0.9867827, tolerance = 1e-5)
+expect_equal(p1$estimate[2], 0.9867827, tolerance = 1e-5)
+expect_equal(p2$estimate[2], 0.9867827, tolerance = 1e-5)
+expect_equal(p3$estimate[2], 0.9867827, tolerance = 1e-5)